Move stubs under dedicated namespace
