+++ stubs/bot_stub.py
