+++ stubs/openai/__init__.py
